Add Rcpp Bindings for `shape`

+++ libtiledbsoma/test/unit_soma_reader.cc
@@ -179,9 +179,9 @@ TEST_CASE("SOMAReader: nnz") {
         uint64_t nnz = sr->nnz();
         REQUIRE(nnz == expected_nnz);
 
-        std::vector<uint64_t> shape = sr->shape();
+        std::vector<int64_t> shape = sr->shape();
         REQUIRE(shape.size() == 1);
-        REQUIRE(shape[0] == (uint64_t)std::numeric_limits<int64_t>::max() + 1);
+        REQUIRE(shape[0] == std::numeric_limits<int64_t>::max() + 1);
     }
 }
